Adding cleaning of evaluated notebooks to  by default.  (#93)
* Adding cleaning of evaluated notebooks to  by default. Use  to see files that would be removed

* Cleaning up logging

* Cleaning up string formatter

+++ scripts/nbsite_cleandisthtml.py
@@ -11,12 +11,9 @@ import shutil
 import sys
 
 # if only someone had made a way to handle parameters
-htmldir = os.path.abspath(sys.argv[1])
-
-if not (len(sys.argv)>2 and sys.argv[2] == 'take_a_chance'):
-    print("This quickly thrown together script will delete things from your computer; you should check it yourself before you run it. Once you've done so, pass 'take_a_chance' as the second argument.")
-    sys.exit(1)
-
+output = sys.argv[1]
+htmldir = os.path.abspath(output)
+dry_run =  not (len(sys.argv)>2 and sys.argv[2] == 'take_a_chance')
 
 def IGetFiles(d):
     for thing in os.scandir(d):
@@ -26,25 +23,39 @@ def IGetFiles(d):
             yield thing.path
 
 # I think it's ok to assume these exist for a sphinx site...
+if dry_run:
+    print("This is just a dry-run of removing files from:", htmldir)
+else:
+    print("Removing files from:", htmldir)
 
 # (.doctrees in build folder by default only for sphinx<1.8)
 for folder in (".doctrees", "_sources"):
     d = os.path.join(htmldir,folder)
     try:
-        print("removing %s"%d)
-        shutil.rmtree(d)
+        if dry_run:
+            print("would remove", folder)
+        else:
+            print("removing", folder)
+            shutil.rmtree(d)
     except:
         pass
 
 for file_ in ("objects.inv",):
     f = os.path.join(htmldir,file_)
     try:
-        print("removing %s"%f)
-        os.remove(f)
+        if dry_run:
+            print("would remove", file_)
+        else:
+            print("removing", file_)
+            os.remove(f)
     except:
         pass
 
 for path in IGetFiles(htmldir):
     if os.path.splitext(path)[1].lower() == '.ipynb':
-        print("removing %s"%path)
-        os.remove(path)
+        name = path.split(output)[-1]
+        if dry_run:
+            print("would remove", name)
+        else:
+            print("removing", name)
+            os.remove(path)
